added a target dummy prefab with healthbar
