+++ Assets/Materials/TargetDummyMaterial.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: aa3784c3f1f10e64da26d3c2d638e020
+timeCreated: 1507657619
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
