+++ Assets/Resources/Prefabs/Enemies.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 263bbd40a4467814e9d4fee36a9f1d35
+folderAsset: yes
+timeCreated: 1507657519
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
