+++ Assets/Resources/Prefabs/Enemies/TargetDummy.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 23ea31238ecdc984eafb5cd6590fbcde
+timeCreated: 1507657651
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
