+++ Assets/Scripts/CombatScripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 1851bdf97aa7f8d478184ed5da20f1c5
+folderAsset: yes
+timeCreated: 1507660980
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
