+++ Assets/Scripts/CombatScripts/Enemies/LookToCamera.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class LookToCamera : MonoBehaviour {
+
+    public Camera targetCamera;
+	// Update is called once per frame
+	void Update ()
+    {
+        transform.LookAt(transform.position +
+            targetCamera.transform.rotation * Vector3.back,
+            targetCamera.transform.rotation * Vector3.up);
+	}
+}
