+++ Assets/Scripts/CombatScripts/Enemies/LookToCamera.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0aa4e62c4ae2327469b0e42a0226fc49
+timeCreated: 1507664431
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
