+++ Assets/Scripts/CombatScripts/HealthBarTestScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: cc6389950520dec41b54fb3ab0c8e352
+timeCreated: 1507661737
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
