+++ Assets/Scripts/CombatScripts/HealthScript.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class HealthScript : MonoBehaviour
+{
+
+    public float health = 100f;
+
+    public void removeHealth(float amount)
+    {
+        health -= amount;
+        if (health <= 0)
+        {
+            Destroy(this.gameObject);
+        }
+    }
+}
