+++ Assets/Scripts/CombatScripts/HealthScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c288bf69abf2a4b47933cbeac92ab481
+timeCreated: 1507658375
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
