+++ Assets/Scripts/CombatScripts/MeleeScript.cs
@@ -1,0 +1,18 @@
+﻿//using System.Collections;
+//using System.Collections.Generic;
+//using UnityEngine;
+
+//public class MeleeScript : MonoBehaviour {
+
+//    public float enemyDamage = 25f;
+	
+
+//        void OnTriggerEnter(Collider other)
+//        {
+//            if (other.collider.name == "Enemy")
+//            {
+//                collision.gameObject.GetComponent<EnemyHealth>().onDammaged.Invoke(damage, this.transform);
+//            }
+//        }
+    
+//}
