+++ Assets/Scripts/CombatScripts/MeleeScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6c3a255607769c04c8c6137be14b6af9
+timeCreated: 1507658585
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
