+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Environment
   - Pickup
   - HidingSpot
+  - Enemy
   layers:
   - Default
   - TransparentFX
